third_party/WebKit/LayoutTests/http/tests/inspector/network/cached-resource-destroyed-moved-to-storage.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
var image;
function loadFirstImage() {
    image = new Image();
    image.onload = firstImageLoaded;
    document.body.appendChild(image);
    image.src = "resources/resource.php?type=image&random=1&size=400";
}

function firstImageLoaded()
{
    console.log("Done1.");
}

function loadSecondImage() {
    image.onload = secondImageLoaded;
    image.src = "resources/resource.php?type=image&random=1&size=200";
}

function secondImageLoaded()
{
    console.log("Done2.");
}

function test()
{
    var imageRequest;
    InspectorTest.recordNetwork();
    InspectorTest.addConsoleSniffer(step2);
    InspectorTest.evaluateInPage("loadFirstImage()");

    function step2()
    {
        imageRequest = InspectorTest.networkRequests().pop();
        imageRequest.requestContent().then(step3);
    }

    var originalContentLength;
    function step3()
    {
        InspectorTest.addResult(imageRequest.url());
        InspectorTest.addResult("request.type: " + imageRequest.resourceType());
        InspectorTest.addResult("request.content.length after requesting content: " + imageRequest.content.length);
        originalContentLength = imageRequest.content.length;
        InspectorTest.assertTrue(imageRequest.content.length > 0, "No content before destroying CachedResource.");

        InspectorTest.addResult("Releasing cached resource.");
        // Loading another image to the same image element so that the original image cached resource is released.
        InspectorTest.addConsoleSniffer(step4);
        InspectorTest.evaluateInPage("loadSecondImage()");
    }

    function step4(msg)
    {
        // Disable-enable cache to force MemoryCache::evictResources().
        InspectorTest.NetworkAgent.setCacheDisabled(true, step5);
    }

    function step5(msg)
    {
        InspectorTest.NetworkAgent.setCacheDisabled(false, step6);
    }

    function step6()
    {
        // Re-request content now that CachedResource should have been destroyed.
        delete imageRequest._content;
        imageRequest.requestContent().then(step7);
    }

    function step7()
    {
        InspectorTest.addResult("request.content.length after requesting content: " + imageRequest.content.length);
        originalContentLength = imageRequest.content.length;
        InspectorTest.assertTrue(imageRequest.content.length === originalContentLength, "Content changed after cached resource was destroyed");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests content is moved from cached resource to resource agent's data storage when cached resource is destroyed.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=92108">Bug 92108</a>
</body>
</html>
