LayoutTests/http/tests/inspector/network/cached-resource-destroyed-moved-to-storage-expected.txt

CONSOLE MESSAGE: line 16: Done1.
CONSOLE MESSAGE: line 26: Done2.
Tests content is moved from cached resource to resource agent's data storage when cached resource is destroyed.

Bug 92108  
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=400
request.type: image
request.content.length after requesting content: 536
Releasing cached resource.
request.content.length after requesting content: 536
